# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}

export interface CreateProjectInput {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface Project {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  completionScore?: number | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  originalFileName?: string | null;
  styleTags: string[];
}

export interface JobStatus {
  id: string;
  projectId: string;
  type: string;
  state: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  error?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ProjectDetail {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  completionScore?: number | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  originalFileName?: string | null;
  styleTags: string[];
  jobs: JobStatus[];
  warnings: string[];
}

export interface UploadResponse {
  projectId: string;
  jobId: string;
  fileName: string;
  status: string;
}

export interface TrackNode {
  id: string;
  name: string;
  type: string;
  /** @nullable */
  parentGroupId?: string | null;
  orderIndex: number;
  muted: boolean;
  solo: boolean;
  frozen: boolean;
  armed: boolean;
  /** @nullable */
  color?: number | null;
  inferredRole: string;
  inferredConfidence: number;
  clipCount: number;
  deviceCount: number;
  automationPoints: number;
  warnings: string[];
}

export interface ArrangementSection {
  id: string;
  label: string;
  startBar: number;
  endBar: number;
  energyScore: number;
  densityScore: number;
  dominantRoles: string[];
  missingElements: string[];
  transitionQuality: number;
}

export interface ProjectGraph {
  projectId: string;
  tempo: number;
  timeSignatureNumerator: number;
  timeSignatureDenominator: number;
  arrangementLength: number;
  tracks: TrackNode[];
  sections: ArrangementSection[];
  parseQuality: number;
  warnings: string[];
  styleTags: string[];
  totalClips: number;
  totalDevices: number;
  returnTrackCount: number;
  masterTrackPresent: boolean;
}

export interface CompletionAction {
  id: string;
  category: string;
  title: string;
  description: string;
  affectedTracks: string[];
  /** @nullable */
  affectedBars?: string | null;
  confidence: number;
  expectedImpact: string;
  rationale: string;
  priority: string;
}

export interface CompletionPlan {
  projectId: string;
  summary: string;
  confidence: number;
  styleTags: string[];
  completionScore: number;
  actions: CompletionAction[];
  weaknesses: string[];
  warnings: string[];
  rationale: string;
  generatedAt: string;
}

export interface ArtifactFile {
  id: string;
  projectId: string;
  type: string;
  fileName: string;
  fileSize: number;
  mimeType: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}

export type UploadAlsFileBody = {
  file?: Blob;
};