# lib/api-zod/src/generated/types/arrangementSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ArrangementSection {
  id: string;
  label: string;
  startBar: number;
  endBar: number;
  energyScore: number;
  densityScore: number;
  dominantRoles: string[];
  missingElements: string[];
  transitionQuality: number;
}