# lib/api-zod/src/generated/types/artifactFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ArtifactFile {
  id: string;
  projectId: string;
  type: string;
  fileName: string;
  fileSize: number;
  mimeType: string;
  createdAt: Date;
  /** @nullable */
  description?: string | null;
}