# lib/api-zod/src/generated/types/completionAction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CompletionAction {
  id: string;
  category: string;
  title: string;
  description: string;
  affectedTracks: string[];
  /** @nullable */
  affectedBars?: string | null;
  confidence: number;
  expectedImpact: string;
  rationale: string;
  priority: string;
}