# lib/api-zod/src/generated/types/completionPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CompletionAction } from "./completionAction";

export interface CompletionPlan {
  projectId: string;
  summary: string;
  confidence: number;
  styleTags: string[];
  completionScore: number;
  actions: CompletionAction[];
  weaknesses: string[];
  warnings: string[];
  rationale: string;
  generatedAt: Date;
}