# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  name: string;
  /** @nullable */
  description?: string | null;
}