# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface JobStatus {
  id: string;
  projectId: string;
  type: string;
  state: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  error?: string | null;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  completedAt?: Date | null;
}