# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from "./jobStatus";

export interface ProjectDetail {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  completionScore?: number | null;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  originalFileName?: string | null;
  styleTags: string[];
  jobs: JobStatus[];
  warnings: string[];
}