# lib/api-zod/src/generated/types/projectGraph.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ArrangementSection } from "./arrangementSection";
import type { TrackNode } from "./trackNode";

export interface ProjectGraph {
  projectId: string;
  tempo: number;
  timeSignatureNumerator: number;
  timeSignatureDenominator: number;
  arrangementLength: number;
  tracks: TrackNode[];
  sections: ArrangementSection[];
  parseQuality: number;
  warnings: string[];
  styleTags: string[];
  totalClips: number;
  totalDevices: number;
  returnTrackCount: number;
  masterTrackPresent: boolean;
}