# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ableton AI Track Completion Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  projectId: string;
  jobId: string;
  fileName: string;
  status: string;
}